BERTAUX/messagerie/requirements.txt
pycryptodome
msgpack
python-prctl
